+++ CHANGELOG.md
@@ -93,6 +93,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Removed `PlayerStorage#tick` and `PlayerStorage#shutdown` methods, these methods shouldn't be in api module.
 - (API) Removed `Registries#BLOCK_STATE_DATA`. This registry is moved to `InternalRegistries`.
 - (API) Removed `Registries#ITEM_DATA`. This registry is moved to `InternalRegistries`.
+- (API) Removed `VoxelShapes#buildStairShape` method, we now have accurate collision shape data dumped from BDS.
 - Removed `Extension#afterServerStarted` method.
 - Removed `org.allaymc.server.datastruct.collections.nb.*`, we now use the implementations provided by JCTools. Consider using `NonBlockingHashMap`
   and `NonBlockingHashMapLong` if your plugins use these classes.